+++ ckanext/gobar_theme/helpers/__init__.py
@@ -2,3 +2,4 @@
 from .helpers import *
 from .organizations import *
 from .groups import *
+from .licenses import *
